Поделили js код на модули
