+++ js/create-messages.js
@@ -1,0 +1,6 @@
+import { getRandomNumber, getArrayElement } from './util';
+import { COMMENTS } from './data';
+
+const createMessage = () => Array.from({length: getRandomNumber(1, 2)}, () => getArrayElement(COMMENTS)).join(' ');
+
+export {createMessage};
